feat: add article api

+++ gmonad/controllers/event.go
@@ -40,13 +40,13 @@ func CreateEvent(c *gin.Context) {
 		return
 	}
 
-	userId, ok := uid.(int)
+	userId, ok := uid.(uint)
 	if !ok {
 		utils.ErrorResponse(c, http.StatusUnauthorized, "unauthorized", nil)
 		return
 	}
 
-	event.UserId = uint(userId)
+	event.UserId = userId
 	// 创建数据库记录
 	if err := event.Create(); err != nil {
 		utils.ErrorResponse(c, http.StatusInternalServerError, err.Error(), nil)
